llamacpp/completion_server/pyproject.toml
[tool.poetry]
name = "completion-server"
version = "0.1.0"
description = ""
authors = ["Louis Brunet <lbrunet@neoxia.deloitte.fr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.3"
pydantic = "^2.9.2"
uvicorn = "^0.32.0"
llama-cpp-python = "^0.3.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
